@@ -37,27 +37,23 @@
         "sass-loader": "^13.2.0",
         "tailwindcss": "^3.1.0",
         "vite": "^3.0.0",
-        "vue-loader": "^15.10.0",
-        "vue-stripe-elements-plus": "^1.0.3",
-        "vue-template-compiler": "^2.7.13"
+        "vue-loader": "^16.8.3",
+        "vue-stripe-elements-plus": "^1.0.3"
     },
     "dependencies": {
         "air-datepicker": "^3.3.0",
         "bootstrap": "^5.2.2",
         "dotenv": "^16.0.3",
-        "gmap-vue": "^3.5.4",
         "google-maps-api-loader": "^1.1.1",
         "moment": "^2.29.4",
-        "tiptap-vuetify": "^2.24.0",
-        "vue": "^2.7.13",
+        "vue": "^3.4.33",
         "vue-awesome-countdown": "^1.1.4",
         "vue-axios": "^2.1.5",
         "vue-infinite-scroll": "^2.0.2",
         "vue-read-more": "^1.1.1",
-        "vue-router": "^3.6.5",
+        "vue-router": "^4.4.0",
         "vue-share-buttons": "^1.1.1",
         "vue2-google-maps": "^0.10.7",
-        "vuetify": "^2.6.12",
-        "vuex": "^3.6.2"
+        "vuetify": "^3.6.13"
     }
 }
