@@ -1,18 +1,18 @@
 "use strict";
 (self["webpackChunk"] = self["webpackChunk"] || []).push([["resources_js_components_lecture_ShowComments_vue"],{
 
-/***/ "./node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!./node_modules/vue-loader/lib/index.js??vue-loader-options!./resources/js/components/lecture/ShowComments.vue?vue&type=script&lang=js&":
-/*!***************************************************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!./node_modules/vue-loader/lib/index.js??vue-loader-options!./resources/js/components/lecture/ShowComments.vue?vue&type=script&lang=js& ***!
-  \***************************************************************************************************************************************************************************************************************/
+/***/ "./node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!./node_modules/vue-loader/dist/index.js??ruleSet[0].use[0]!./resources/js/components/lecture/ShowComments.vue?vue&type=script&lang=js":
+/*!**************************************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!./node_modules/vue-loader/dist/index.js??ruleSet[0].use[0]!./resources/js/components/lecture/ShowComments.vue?vue&type=script&lang=js ***!
+  \**************************************************************************************************************************************************************************************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 __webpack_require__.r(__webpack_exports__);
 /* harmony export */ __webpack_require__.d(__webpack_exports__, {
 /* harmony export */   "default": () => (__WEBPACK_DEFAULT_EXPORT__)
 /* harmony export */ });
-/* harmony import */ var vuex__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! vuex */ "./node_modules/vuex/dist/vuex.esm.js");
-/* harmony import */ var tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! tiptap-vuetify */ "./node_modules/tiptap-vuetify/dist/bundle-esm.js");
+Object(function webpackMissingModule() { var e = new Error("Cannot find module 'vuex'"); e.code = 'MODULE_NOT_FOUND'; throw e; }());
+Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }());
 /* harmony import */ var _services_auth_auth__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../services/auth/auth */ "./resources/js/services/auth/auth.js");
 /* harmony import */ var _loader_ExportLoading__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../loader/ExportLoading */ "./resources/js/components/loader/ExportLoading.vue");
 function ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); enumerableOnly && (symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; })), keys.push.apply(keys, symbols); } return keys; }
@@ -35,18 +35,18 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
         comment: ''
       },
       editField: false,
-      extensions: [tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.History, tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.Blockquote, tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.Link, tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.Underline, tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.Strike, tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.Italic, tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.ListItem, tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.BulletList, tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.OrderedList, [tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.Heading, {
+      extensions: [Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }()), Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }()), Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }()), Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }()), Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }()), Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }()), Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }()), Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }()), Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }()), [Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }()), {
         options: {
           levels: [1, 2, 3]
         }
-      }], tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.Bold, tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.Code, tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.HorizontalRule, tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.Paragraph, tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.HardBreak]
+      }], Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }()), Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }()), Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }()), Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }()), Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }())]
     };
   },
   components: {
     ExportLoading: _loader_ExportLoading__WEBPACK_IMPORTED_MODULE_2__["default"],
-    TiptapVuetify: tiptap_vuetify__WEBPACK_IMPORTED_MODULE_0__.TiptapVuetify
+    TiptapVuetify: Object(function webpackMissingModule() { var e = new Error("Cannot find module 'tiptap-vuetify'"); e.code = 'MODULE_NOT_FOUND'; throw e; }())
   },
-  computed: (0,vuex__WEBPACK_IMPORTED_MODULE_3__.mapState)({
+  computed: Object(function webpackMissingModule() { var e = new Error("Cannot find module 'vuex'"); e.code = 'MODULE_NOT_FOUND'; throw e; }())({
     comments: function comments(state) {
       return state.comment.comments;
     },
@@ -64,7 +64,7 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
     }
   }),
   mounted: function mounted() {},
-  methods: _objectSpread(_objectSpread({}, (0,vuex__WEBPACK_IMPORTED_MODULE_3__.mapActions)('comment', ['getComments', 'addComment', 'updateComment', 'exportComments'])), {}, {
+  methods: _objectSpread(_objectSpread({}, Object(function webpackMissingModule() { var e = new Error("Cannot find module 'vuex'"); e.code = 'MODULE_NOT_FOUND'; throw e; }())('comment', ['getComments', 'addComment', 'updateComment', 'exportComments'])), {}, {
     loadMore: function loadMore() {
       this.getComments(this.$route.params.id);
     },
@@ -76,164 +76,222 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
 
 /***/ }),
 
-/***/ "./node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!./node_modules/vue-loader/lib/loaders/templateLoader.js??ruleSet[1].rules[2]!./node_modules/vue-loader/lib/index.js??vue-loader-options!./resources/js/components/lecture/ShowComments.vue?vue&type=template&id=7516de26&scoped=true&":
-/*!**************************************************************************************************************************************************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!./node_modules/vue-loader/lib/loaders/templateLoader.js??ruleSet[1].rules[2]!./node_modules/vue-loader/lib/index.js??vue-loader-options!./resources/js/components/lecture/ShowComments.vue?vue&type=template&id=7516de26&scoped=true& ***!
-  \**************************************************************************************************************************************************************************************************************************************************************************************************************/
+/***/ "./node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!./node_modules/vue-loader/dist/templateLoader.js??ruleSet[1].rules[2]!./node_modules/vue-loader/dist/index.js??ruleSet[0].use[0]!./resources/js/components/lecture/ShowComments.vue?vue&type=template&id=7516de26":
+/*!******************************************************************************************************************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!./node_modules/vue-loader/dist/templateLoader.js??ruleSet[1].rules[2]!./node_modules/vue-loader/dist/index.js??ruleSet[0].use[0]!./resources/js/components/lecture/ShowComments.vue?vue&type=template&id=7516de26 ***!
+  \******************************************************************************************************************************************************************************************************************************************************************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 __webpack_require__.r(__webpack_exports__);
 /* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   "render": () => (/* binding */ render),
-/* harmony export */   "staticRenderFns": () => (/* binding */ staticRenderFns)
+/* harmony export */   "render": () => (/* binding */ render)
 /* harmony export */ });
-var render = function render() {
-  var _vm = this,
-    _c = _vm._self._c;
-  return _c("div", {
-    staticClass: "mt-3"
-  }, [_c("v-dialog", {
-    attrs: {
-      width: "500"
-    },
-    model: {
-      value: _vm.dialog,
-      callback: function callback($$v) {
-        _vm.dialog = $$v;
-      },
-      expression: "dialog"
-    }
-  }, [_c("export-loading")], 1), _vm._v(" "), _c("v-layout", {
-    attrs: {
-      "justify-center": ""
-    }
-  }, [_c("v-card", {
-    attrs: {
-      width: "50%"
-    }
-  }, [_c("v-card-title", {
-    staticClass: "text-h4"
-  }, [_vm._v("Comments " + _vm._s(_vm.quantity))]), _c("hr", {
-    staticClass: "marg-auto",
-    attrs: {
-      width: "90%"
-    }
-  }), _vm._v(" "), _c("v-card-subtitle", {
-    directives: [{
-      name: "show",
-      rawName: "v-show",
-      value: _vm.is_admin,
-      expression: "is_admin"
-    }]
-  }, [_c("v-btn", {
-    attrs: {
-      color: "blue",
-      text: ""
-    },
-    on: {
-      click: function click($event) {
-        _vm.exportComments(_vm.$route.params.id);
-        _vm.dialog = true;
-      }
-    }
-  }, [_vm._v("Export comments")])], 1), _vm._v(" "), _c("v-card-text", [_c("v-form", {
-    on: {
-      submit: function submit($event) {
-        $event.preventDefault();
-        return _vm.addComment(_vm.newComment);
-      }
-    }
-  }, [_c("tiptap-vuetify", {
-    staticClass: "mb-2",
-    attrs: {
-      extensions: _vm.extensions
-    },
-    model: {
-      value: _vm.newComment.comment,
-      callback: function callback($$v) {
-        _vm.$set(_vm.newComment, "comment", $$v);
-      },
-      expression: "newComment.comment"
-    }
-  }), _vm._v(" "), _c("v-btn", {
-    attrs: {
-      color: "secondary",
-      type: "submit"
-    }
-  }, [_vm._v("Send")])], 1)], 1), _vm._v(" "), _c("v-card-text", {
-    staticClass: "text--primary"
-  }, [_c("div", {
-    directives: [{
-      name: "infinite-scroll",
-      rawName: "v-infinite-scroll",
-      value: _vm.loadMore,
-      expression: "loadMore"
-    }],
-    attrs: {
-      "infinite-scroll-disabled": "busy",
-      "infinite-scroll-distance": "limit"
-    }
-  }, [_vm._l(_vm.comments, function (comment) {
-    return _c("v-card", {
-      key: comment.id,
-      staticClass: "mb-2"
-    }, [_c("v-card-title", {
-      staticClass: "text-h6"
-    }, [_vm._v(_vm._s(comment.firstname) + " " + _vm._s(comment.lastname))]), _vm._v(" "), _c("v-card-subtitle", [_vm._v(_vm._s(new Date(comment.created_at).toLocaleDateString("en-us", {
-      weekday: "long",
-      year: "numeric",
-      month: "short",
-      day: "numeric"
-    })))]), _vm._v(" "), _c("v-card-text", {
-      staticClass: "text--primary",
-      domProps: {
-        innerHTML: _vm._s(comment.comment)
-      }
-    }), _vm._v(" "), comment.user_id === _vm.user && Math.round((new Date() - new Date(comment.created_at)) % 86400000 % 3600000 / 60000) < 10 ? _c("v-card-actions", [_c("v-btn", {
-      attrs: {
-        text: "",
-        align: "right"
-      },
-      on: {
-        click: function click($event) {
-          _vm.editField = !_vm.editField;
-        }
-      }
-    }, [_vm._v("Edit")]), _vm._v(" "), _vm.editField ? _c("v-card-text", [_c("v-form", {
-      on: {
-        submit: function submit($event) {
-          $event.preventDefault();
-          return _vm.updateComment(comment);
-        }
-      }
-    }, [_c("tiptap-vuetify", {
-      staticClass: "mb-2",
-      attrs: {
-        extensions: _vm.extensions
-      },
-      model: {
-        value: comment.comment,
-        callback: function callback($$v) {
-          _vm.$set(comment, "comment", $$v);
-        },
-        expression: "comment.comment"
-      }
-    }), _vm._v(" "), _c("v-btn", {
-      attrs: {
-        text: "",
-        type: "submit"
-      }
-    }, [_vm._v("Save")])], 1)], 1) : _vm._e()], 1) : _vm._e()], 1);
-  }), _vm._v(" "), _vm.loading ? _c("v-progress-circular", {
-    attrs: {
-      indeterminate: "",
-      color: "primary"
-    }
-  }) : _vm._e()], 2)])], 1)], 1)], 1);
-};
-var staticRenderFns = [];
-render._withStripped = true;
+/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! vue */ "./node_modules/vue/dist/vue.esm-bundler.js");
 
+var _hoisted_1 = {
+  "class": "mt-3"
+};
+var _hoisted_2 = /*#__PURE__*/(0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementVNode)("hr", {
+  width: "90%",
+  "class": "marg-auto"
+}, null, -1 /* HOISTED */);
+var _hoisted_3 = {
+  "infinite-scroll-disabled": "busy",
+  "infinite-scroll-distance": "limit"
+};
+function render(_ctx, _cache, $props, $setup, $data, $options) {
+  var _component_export_loading = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("export-loading");
+  var _component_v_dialog = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("v-dialog");
+  var _component_v_card_title = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("v-card-title");
+  var _component_v_btn = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("v-btn");
+  var _component_v_card_subtitle = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("v-card-subtitle");
+  var _component_tiptap_vuetify = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("tiptap-vuetify");
+  var _component_v_form = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("v-form");
+  var _component_v_card_text = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("v-card-text");
+  var _component_v_card_actions = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("v-card-actions");
+  var _component_v_card = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("v-card");
+  var _component_v_progress_circular = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("v-progress-circular");
+  var _component_v_layout = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveComponent)("v-layout");
+  var _directive_infinite_scroll = (0,vue__WEBPACK_IMPORTED_MODULE_0__.resolveDirective)("infinite-scroll");
+  return (0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("div", _hoisted_1, [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_dialog, {
+    modelValue: $data.dialog,
+    "onUpdate:modelValue": _cache[0] || (_cache[0] = function ($event) {
+      return $data.dialog = $event;
+    }),
+    width: "500"
+  }, {
+    "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+      return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_export_loading)];
+    }),
+    _: 1 /* STABLE */
+  }, 8 /* PROPS */, ["modelValue"]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_layout, {
+    "justify-center": ""
+  }, {
+    "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+      return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_card, {
+        width: "50%"
+      }, {
+        "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+          return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_card_title, {
+            "class": "text-h4"
+          }, {
+            "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+              return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createTextVNode)("Comments " + (0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)(_ctx.quantity), 1 /* TEXT */)];
+            }),
+
+            _: 1 /* STABLE */
+          }), _hoisted_2, (0,vue__WEBPACK_IMPORTED_MODULE_0__.withDirectives)((0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_card_subtitle, null, {
+            "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+              return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_btn, {
+                onClick: _cache[1] || (_cache[1] = function ($event) {
+                  _ctx.exportComments(_ctx.$route.params.id);
+                  $data.dialog = true;
+                }),
+                color: "blue",
+                text: ""
+              }, {
+                "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+                  return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createTextVNode)("Export comments")];
+                }),
+                _: 1 /* STABLE */
+              })];
+            }),
+
+            _: 1 /* STABLE */
+          }, 512 /* NEED_PATCH */), [[vue__WEBPACK_IMPORTED_MODULE_0__.vShow, $data.is_admin]]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_card_text, null, {
+            "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+              return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_form, {
+                onSubmit: _cache[3] || (_cache[3] = (0,vue__WEBPACK_IMPORTED_MODULE_0__.withModifiers)(function ($event) {
+                  return _ctx.addComment($data.newComment);
+                }, ["prevent"]))
+              }, {
+                "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+                  return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_tiptap_vuetify, {
+                    modelValue: $data.newComment.comment,
+                    "onUpdate:modelValue": _cache[2] || (_cache[2] = function ($event) {
+                      return $data.newComment.comment = $event;
+                    }),
+                    extensions: $data.extensions,
+                    "class": "mb-2"
+                  }, null, 8 /* PROPS */, ["modelValue", "extensions"]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_btn, {
+                    color: "secondary",
+                    type: "submit"
+                  }, {
+                    "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+                      return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createTextVNode)("Send")];
+                    }),
+                    _: 1 /* STABLE */
+                  })];
+                }),
+
+                _: 1 /* STABLE */
+              })];
+            }),
+
+            _: 1 /* STABLE */
+          }), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_card_text, {
+            "class": "text--primary"
+          }, {
+            "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+              return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.withDirectives)(((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)("div", _hoisted_3, [((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(true), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createElementBlock)(vue__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, (0,vue__WEBPACK_IMPORTED_MODULE_0__.renderList)(_ctx.comments, function (comment) {
+                return (0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createBlock)(_component_v_card, {
+                  key: comment.id,
+                  "class": "mb-2"
+                }, {
+                  "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+                    return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_card_title, {
+                      "class": "text-h6"
+                    }, {
+                      "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+                        return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createTextVNode)((0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)(comment.firstname) + " " + (0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)(comment.lastname), 1 /* TEXT */)];
+                      }),
+
+                      _: 2 /* DYNAMIC */
+                    }, 1024 /* DYNAMIC_SLOTS */), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_card_subtitle, null, {
+                      "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+                        return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createTextVNode)((0,vue__WEBPACK_IMPORTED_MODULE_0__.toDisplayString)(new Date(comment.created_at).toLocaleDateString('en-us', {
+                          weekday: "long",
+                          year: "numeric",
+                          month: "short",
+                          day: "numeric"
+                        })), 1 /* TEXT */)];
+                      }),
+
+                      _: 2 /* DYNAMIC */
+                    }, 1024 /* DYNAMIC_SLOTS */), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_card_text, {
+                      "class": "text--primary",
+                      innerHTML: comment.comment
+                    }, null, 8 /* PROPS */, ["innerHTML"]), comment.user_id === $data.user && Math.round((new Date() - new Date(comment.created_at)) % 86400000 % 3600000 / 60000) < 10 ? ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createBlock)(_component_v_card_actions, {
+                      key: 0
+                    }, {
+                      "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+                        return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_btn, {
+                          text: "",
+                          align: "right",
+                          onClick: _cache[4] || (_cache[4] = function ($event) {
+                            return $data.editField = !$data.editField;
+                          })
+                        }, {
+                          "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+                            return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createTextVNode)("Edit")];
+                          }),
+                          _: 1 /* STABLE */
+                        }), $data.editField ? ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createBlock)(_component_v_card_text, {
+                          key: 0
+                        }, {
+                          "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+                            return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_form, {
+                              onSubmit: (0,vue__WEBPACK_IMPORTED_MODULE_0__.withModifiers)(function ($event) {
+                                return _ctx.updateComment(comment);
+                              }, ["prevent"])
+                            }, {
+                              "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+                                return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_tiptap_vuetify, {
+                                  modelValue: comment.comment,
+                                  "onUpdate:modelValue": function onUpdateModelValue($event) {
+                                    return comment.comment = $event;
+                                  },
+                                  extensions: $data.extensions,
+                                  "class": "mb-2"
+                                }, null, 8 /* PROPS */, ["modelValue", "onUpdate:modelValue", "extensions"]), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createVNode)(_component_v_btn, {
+                                  text: "",
+                                  type: "submit"
+                                }, {
+                                  "default": (0,vue__WEBPACK_IMPORTED_MODULE_0__.withCtx)(function () {
+                                    return [(0,vue__WEBPACK_IMPORTED_MODULE_0__.createTextVNode)("Save")];
+                                  }),
+                                  _: 1 /* STABLE */
+                                })];
+                              }),
+
+                              _: 2 /* DYNAMIC */
+                            }, 1032 /* PROPS, DYNAMIC_SLOTS */, ["onSubmit"])];
+                          }),
+                          _: 2 /* DYNAMIC */
+                        }, 1024 /* DYNAMIC_SLOTS */)) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true)];
+                      }),
+                      _: 2 /* DYNAMIC */
+                    }, 1024 /* DYNAMIC_SLOTS */)) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true)];
+                  }),
+                  _: 2 /* DYNAMIC */
+                }, 1024 /* DYNAMIC_SLOTS */);
+              }), 128 /* KEYED_FRAGMENT */)), _ctx.loading ? ((0,vue__WEBPACK_IMPORTED_MODULE_0__.openBlock)(), (0,vue__WEBPACK_IMPORTED_MODULE_0__.createBlock)(_component_v_progress_circular, {
+                key: 0,
+                indeterminate: "",
+                color: "primary"
+              })) : (0,vue__WEBPACK_IMPORTED_MODULE_0__.createCommentVNode)("v-if", true)])), [[_directive_infinite_scroll, $options.loadMore]])];
+            }),
+            _: 1 /* STABLE */
+          })];
+        }),
+
+        _: 1 /* STABLE */
+      })];
+    }),
+
+    _: 1 /* STABLE */
+  })]);
+}
 
 /***/ }),
 
@@ -247,61 +305,49 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony export */ __webpack_require__.d(__webpack_exports__, {
 /* harmony export */   "default": () => (__WEBPACK_DEFAULT_EXPORT__)
 /* harmony export */ });
-/* harmony import */ var _ShowComments_vue_vue_type_template_id_7516de26_scoped_true___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./ShowComments.vue?vue&type=template&id=7516de26&scoped=true& */ "./resources/js/components/lecture/ShowComments.vue?vue&type=template&id=7516de26&scoped=true&");
-/* harmony import */ var _ShowComments_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./ShowComments.vue?vue&type=script&lang=js& */ "./resources/js/components/lecture/ShowComments.vue?vue&type=script&lang=js&");
-/* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! !../../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ "./node_modules/vue-loader/lib/runtime/componentNormalizer.js");
+/* harmony import */ var _ShowComments_vue_vue_type_template_id_7516de26__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./ShowComments.vue?vue&type=template&id=7516de26 */ "./resources/js/components/lecture/ShowComments.vue?vue&type=template&id=7516de26");
+/* harmony import */ var _ShowComments_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./ShowComments.vue?vue&type=script&lang=js */ "./resources/js/components/lecture/ShowComments.vue?vue&type=script&lang=js");
+/* harmony import */ var _Users_ksumasalitina_personal_projects_conferences_laravel_nova_project_node_modules_vue_loader_dist_exportHelper_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./node_modules/vue-loader/dist/exportHelper.js */ "./node_modules/vue-loader/dist/exportHelper.js");
 
 
 
 
-
-/* normalize component */
 ;
-var component = (0,_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__["default"])(
-  _ShowComments_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__["default"],
-  _ShowComments_vue_vue_type_template_id_7516de26_scoped_true___WEBPACK_IMPORTED_MODULE_0__.render,
-  _ShowComments_vue_vue_type_template_id_7516de26_scoped_true___WEBPACK_IMPORTED_MODULE_0__.staticRenderFns,
-  false,
-  null,
-  "7516de26",
-  null
-  
-)
-
+const __exports__ = /*#__PURE__*/(0,_Users_ksumasalitina_personal_projects_conferences_laravel_nova_project_node_modules_vue_loader_dist_exportHelper_js__WEBPACK_IMPORTED_MODULE_2__["default"])(_ShowComments_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__["default"], [['render',_ShowComments_vue_vue_type_template_id_7516de26__WEBPACK_IMPORTED_MODULE_0__.render],['__file',"resources/js/components/lecture/ShowComments.vue"]])
 /* hot reload */
-if (false) { var api; }
-component.options.__file = "resources/js/components/lecture/ShowComments.vue"
-/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (component.exports);
+if (false) {}
+
+
+/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (__exports__);
 
 /***/ }),
 
-/***/ "./resources/js/components/lecture/ShowComments.vue?vue&type=script&lang=js&":
-/*!***********************************************************************************!*\
-  !*** ./resources/js/components/lecture/ShowComments.vue?vue&type=script&lang=js& ***!
-  \***********************************************************************************/
+/***/ "./resources/js/components/lecture/ShowComments.vue?vue&type=script&lang=js":
+/*!**********************************************************************************!*\
+  !*** ./resources/js/components/lecture/ShowComments.vue?vue&type=script&lang=js ***!
+  \**********************************************************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 __webpack_require__.r(__webpack_exports__);
 /* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   "default": () => (__WEBPACK_DEFAULT_EXPORT__)
+/* harmony export */   "default": () => (/* reexport safe */ _node_modules_babel_loader_lib_index_js_clonedRuleSet_5_use_0_node_modules_vue_loader_dist_index_js_ruleSet_0_use_0_ShowComments_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_0__["default"])
 /* harmony export */ });
-/* harmony import */ var _node_modules_babel_loader_lib_index_js_clonedRuleSet_5_use_0_node_modules_vue_loader_lib_index_js_vue_loader_options_ShowComments_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!../../../../node_modules/vue-loader/lib/index.js??vue-loader-options!./ShowComments.vue?vue&type=script&lang=js& */ "./node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!./node_modules/vue-loader/lib/index.js??vue-loader-options!./resources/js/components/lecture/ShowComments.vue?vue&type=script&lang=js&");
- /* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (_node_modules_babel_loader_lib_index_js_clonedRuleSet_5_use_0_node_modules_vue_loader_lib_index_js_vue_loader_options_ShowComments_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__["default"]); 
+/* harmony import */ var _node_modules_babel_loader_lib_index_js_clonedRuleSet_5_use_0_node_modules_vue_loader_dist_index_js_ruleSet_0_use_0_ShowComments_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!../../../../node_modules/vue-loader/dist/index.js??ruleSet[0].use[0]!./ShowComments.vue?vue&type=script&lang=js */ "./node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!./node_modules/vue-loader/dist/index.js??ruleSet[0].use[0]!./resources/js/components/lecture/ShowComments.vue?vue&type=script&lang=js");
+ 
 
 /***/ }),
 
-/***/ "./resources/js/components/lecture/ShowComments.vue?vue&type=template&id=7516de26&scoped=true&":
-/*!*****************************************************************************************************!*\
-  !*** ./resources/js/components/lecture/ShowComments.vue?vue&type=template&id=7516de26&scoped=true& ***!
-  \*****************************************************************************************************/
+/***/ "./resources/js/components/lecture/ShowComments.vue?vue&type=template&id=7516de26":
+/*!****************************************************************************************!*\
+  !*** ./resources/js/components/lecture/ShowComments.vue?vue&type=template&id=7516de26 ***!
+  \****************************************************************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 __webpack_require__.r(__webpack_exports__);
 /* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   "render": () => (/* reexport safe */ _node_modules_babel_loader_lib_index_js_clonedRuleSet_5_use_0_node_modules_vue_loader_lib_loaders_templateLoader_js_ruleSet_1_rules_2_node_modules_vue_loader_lib_index_js_vue_loader_options_ShowComments_vue_vue_type_template_id_7516de26_scoped_true___WEBPACK_IMPORTED_MODULE_0__.render),
-/* harmony export */   "staticRenderFns": () => (/* reexport safe */ _node_modules_babel_loader_lib_index_js_clonedRuleSet_5_use_0_node_modules_vue_loader_lib_loaders_templateLoader_js_ruleSet_1_rules_2_node_modules_vue_loader_lib_index_js_vue_loader_options_ShowComments_vue_vue_type_template_id_7516de26_scoped_true___WEBPACK_IMPORTED_MODULE_0__.staticRenderFns)
+/* harmony export */   "render": () => (/* reexport safe */ _node_modules_babel_loader_lib_index_js_clonedRuleSet_5_use_0_node_modules_vue_loader_dist_templateLoader_js_ruleSet_1_rules_2_node_modules_vue_loader_dist_index_js_ruleSet_0_use_0_ShowComments_vue_vue_type_template_id_7516de26__WEBPACK_IMPORTED_MODULE_0__.render)
 /* harmony export */ });
-/* harmony import */ var _node_modules_babel_loader_lib_index_js_clonedRuleSet_5_use_0_node_modules_vue_loader_lib_loaders_templateLoader_js_ruleSet_1_rules_2_node_modules_vue_loader_lib_index_js_vue_loader_options_ShowComments_vue_vue_type_template_id_7516de26_scoped_true___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!../../../../node_modules/vue-loader/lib/loaders/templateLoader.js??ruleSet[1].rules[2]!../../../../node_modules/vue-loader/lib/index.js??vue-loader-options!./ShowComments.vue?vue&type=template&id=7516de26&scoped=true& */ "./node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!./node_modules/vue-loader/lib/loaders/templateLoader.js??ruleSet[1].rules[2]!./node_modules/vue-loader/lib/index.js??vue-loader-options!./resources/js/components/lecture/ShowComments.vue?vue&type=template&id=7516de26&scoped=true&");
+/* harmony import */ var _node_modules_babel_loader_lib_index_js_clonedRuleSet_5_use_0_node_modules_vue_loader_dist_templateLoader_js_ruleSet_1_rules_2_node_modules_vue_loader_dist_index_js_ruleSet_0_use_0_ShowComments_vue_vue_type_template_id_7516de26__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!../../../../node_modules/vue-loader/dist/templateLoader.js??ruleSet[1].rules[2]!../../../../node_modules/vue-loader/dist/index.js??ruleSet[0].use[0]!./ShowComments.vue?vue&type=template&id=7516de26 */ "./node_modules/babel-loader/lib/index.js??clonedRuleSet-5.use[0]!./node_modules/vue-loader/dist/templateLoader.js??ruleSet[1].rules[2]!./node_modules/vue-loader/dist/index.js??ruleSet[0].use[0]!./resources/js/components/lecture/ShowComments.vue?vue&type=template&id=7516de26");
 
 
 /***/ })
